// Code/Websites/DanpheEMR/wwwroot/DanpheApp/src/app/pharmacy/report/expiry-stock/phrm-expiry-report-data.dto.ts
export class PHRMExpiryReportData_DTO {
    ItemId: number = null;
    BatchNo: string = null;
    ExpiryDate: string = null;
    SalePrice: number = 0;
    CostPrice: number = 0;
    ItemName: string = null;
    GenericName: string = null;
    AvailableQuantity: number = 0;
    StoreName: string = null;
    SupplierName: string = null;
    TotalCostValue: number = 0;
    TotalSalesValue: number = 0;
}